Move need_type_info too
