+++ compiler/rustc_infer/src/error_reporting/infer/mod.rs
@@ -84,10 +84,10 @@ use crate::traits::{
 mod note_and_explain;
 mod suggest;
 
+pub mod need_type_info;
+pub mod nice_region_error;
 pub mod region;
 pub mod sub_relations;
-
-pub mod nice_region_error;
 
 /// Makes a valid string literal from a string by escaping special characters (" and \),
 /// unless they are already escaped.
